+++ lib/gks/gksquartz.h
@@ -3,13 +3,13 @@ typedef struct
     int state;
     int win;
     gks_display_list_t dl;
-    float width, height;
-    float swidth, sheight;
-    float a, b, c, d;
-    float window[4], viewport[4];
+    double width, height;
+    double swidth, sheight;
+    double a, b, c, d;
+    double window[4], viewport[4];
     CGColorRef rgb[MAX_COLOR];
     int family, capheight;
-    float angle;
+    double angle;
     CGRect rect[MAX_TNR];
  }
 ws_state_list;
